# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tài liệu API cho BTL Nhóm 2",
        "description": "Đây là tài liệu API cho ứng dụng điểm danh/quản lý lịch dạy.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/leaverequests": {
            "get": {
                "tags": [
                    "Leave Requests"
                ],
                "summary": "Lấy DS Đơn nghỉ (Chưa triển khai)",
                "operationId": "getLeaveRequestsList",
                "responses": {
                    "200": {
                        "description": "Chưa triển khai"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Leave Requests"
                ],
                "summary": "Gửi yêu cầu xin nghỉ (Dùng route resource /api/leaverequests)",
                "description": "Route /api/leave-requests (custom) cũng trỏ về đây",
                "operationId": "storeLeaveRequestResource",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "schedule_id",
                                    "reason"
                                ],
                                "properties": {
                                    "schedule_id": {
                                        "description": "ID của lịch dạy muốn nghỉ",
                                        "type": "integer",
                                        "example": 12
                                    },
                                    "reason": {
                                        "description": "Lý do xin nghỉ",
                                        "type": "string",
                                        "example": "Bị ốm"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Gửi yêu cầu thành công",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Chưa đăng nhập"
                    },
                    "422": {
                        "description": "Lỗi validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/leave-requests": {
            "post": {
                "tags": [
                    "Leave Requests"
                ],
                "summary": "Gửi yêu cầu xin nghỉ (Dùng route /api/leave-requests)",
                "description": "Route /api/leaverequests (resource) cũng trỏ về đây",
                "operationId": "storeLeaveRequest",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "schedule_id",
                                    "reason"
                                ],
                                "properties": {
                                    "schedule_id": {
                                        "description": "ID của lịch dạy muốn nghỉ",
                                        "type": "integer",
                                        "example": 12
                                    },
                                    "reason": {
                                        "description": "Lý do xin nghỉ",
                                        "type": "string",
                                        "example": "Bị ốm"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Gửi yêu cầu thành công",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Chưa đăng nhập"
                    },
                    "422": {
                        "description": "Lỗi validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/leaverequests/{leaverequest}": {
            "get": {
                "tags": [
                    "Leave Requests"
                ],
                "summary": "Xem 1 Đơn nghỉ (Chưa triển khai)",
                "operationId": "getLeaveRequestById",
                "parameters": [
                    {
                        "name": "leaverequest",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chưa triển khai"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Leave Requests"
                ],
                "summary": "Cập nhật Đơn nghỉ (Chưa triển khai)",
                "operationId": "updateLeaveRequest",
                "parameters": [
                    {
                        "name": "leaverequest",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chưa triển khai"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Leave Requests"
                ],
                "summary": "Xóa Đơn nghỉ (Chưa triển khai)",
                "operationId": "deleteLeaveRequest",
                "parameters": [
                    {
                        "name": "leaverequest",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chưa triển khai"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{user}/leave-history": {
            "get": {
                "tags": [
                    "Leave Requests"
                ],
                "summary": "Lấy lịch sử xin nghỉ của giáo viên",
                "operationId": "getLeaveHistoryForTeacher",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID của giáo viên",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "leave_request_id": {
                                                "type": "integer"
                                            },
                                            "subject_name": {
                                                "type": "string"
                                            },
                                            "leave_status": {
                                                "type": "string",
                                                "enum": [
                                                    "pending",
                                                    "approved",
                                                    "rejected"
                                                ]
                                            },
                                            "reason": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Sử dụng Bearer token (Sanctum) sau khi đăng nhập",
                "name": "Bearer Token",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}